Fix squiggle in a user's settings json file
I looked at: https://code.visualstudio.com/api/references/contribution-points#contributes.configuration and found the reason for the error squiggle was simply fixed. All you needed to do is allow for the type to be null here. I've already tested it on my computer by changing it on the extension (in %USERPROFILE%\.vscode\extensions\redvanworkshop.explorer-exclude-vscode-extension-1.2.0 ) and the error squiggle immediately disappeared. No need to make it an empty object or anything.